Added parabolic jumping system.

+++ IMoveable.h
@@ -11,7 +11,7 @@ protected:
 public:
 	enum State { idle, walk, attacking, take_dmg, death, transforming, transforming_back, rolling };
 
-	virtual void update(int dt_, Level& level, sf::RenderWindow& window) = 0;
+	virtual void update(float dt_, Level& level, sf::RenderWindow& window) = 0;
 	virtual void draw(sf::RenderWindow& window) = 0;
 	virtual void setdX(double dx_) = 0;
 	virtual void setdY(double dy_) = 0;
